server/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
cassandra-driver==3.25.0
